requirements.txt
Flask==3.1.0
flask_limiter==3.10.1
Flask_Stats==1.0.1
pandas==2.2.3
psycopg2-binary==2.9.10
openpyxl